pyproject.toml
[project]
name = "docx_flattern_tools"
version = "0.1.0"
description = "A tool to convert docx documents to markdown format"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.9"
dependencies = [
    "python-docx>=1.1.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
docx2md = "docx_flattern_tools.main:convert_docx_to_md"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"